R6 class representing a regression model

Contains fields storing data and methods to build, process and visualize a regression model. Currently, this class is designed for CoxPH and GLM regression models.

Public fields

data

a data.table storing modeling data.

recipe

an R formula storing model formula.

f

a string specifying modeling function, default is 'coxph'.

terms

all terms (covariables, i.e. columns) used for building model.

args

other arguments used for building model.

model

a constructed model.

type

model type (class).

result

model result, a object of parameters_model. Can be converted into data.frame with as.data.frame() or data.table::as.data.table().

Methods

Public methods

REGModel$new()

REGModel$plot()

REGModel$print()

Method new()

Build a REGModel object.

Usage

REGModel$new(
  data,
  recipe,
  ...,
  f = c("coxph", "binomial", "gaussian", "Gamma", "inverse.gaussian", "poisson",
    "quasi", "quasibinomial", "quasipoisson"),
  exp = TRUE,
  ci = 0.95
)

Arguments

data

a data.table storing modeling data.

recipe

an R formula or a list with two elements 'x' and 'y', where 'x' is for covariables and 'y' is for label. See example for detail operation.

...

other parameters passing to corresponding regression model function.

f

a length-1 string specifying modeling function, default is 'coxph'. Other options are members of GLM family, see stats::family(). 'binomial' is logistic, and 'gaussian' is linear.

exp

indicating whether or not to exponentiate the the coefficients.

ci

confidence Interval (CI) level. Default to 0.95 (95%). e.g. survival::coxph().

Returns

a REGModel R6 object.

Method plot()

print the REGModel result with default plot methods from parameters package.

Usage

REGModel$plot()

Method print()

print the REGModel object

Usage

REGModel$print(...)

Arguments

...

unused.

Examples

library(survival)
test1 <- data.frame(
  time = c(4, 3, 1, 1, 2, 2, 3),
  status = c(1, 1, 1, 0, 1, 1, 0),
  x = c(0, 2, 1, 1, 1, 0, 0),
  sex = c(0, 0, 0, 0, 1, 1, 1)
)

# --------------
# Build a model
# --------------

# way 1:
mm <- REGModel$new(
  as.data.frame(test1),
  Surv(time, status) ~ x + strata(sex)
)
#> Registered S3 method overwritten by 'parameters':
#>   method                         from      
#>   format.parameters_distribution datawizard
mm
#> ======================
#> A <REGModel> object
#> ======================
#> Parameter | Coefficient |   SE |        95% CI |    z |     p
#> -------------------------------------------------------------
#> x         |        2.23 | 1.83 | [0.45, 11.18] | 0.98 | 0.329
#> 
#> Uncertainty intervals (equal-tailed) and p values (two-tailed) computed using a
#>   Wald z-distribution approximation.
#> ======================
as.data.frame(mm$result)
#>   Parameter Coefficient      SE   CI    CI_low  CI_high         z df_error
#> 1         x    2.230706 1.83448 0.95 0.4450758 11.18022 0.9756088      Inf
#>           p
#> 1 0.3292583

# way 2:
mm2 <- REGModel$new(
  as.data.frame(test1),
  recipe = list(
    x = c("x", "strata(sex)"),
    y = c("time", "status")
  )
)
mm2
#> ======================
#> A <REGModel> object
#> ======================
#> Parameter | Coefficient |   SE |        95% CI |    z |     p
#> -------------------------------------------------------------
#> x         |        2.23 | 1.83 | [0.45, 11.18] | 0.98 | 0.329
#> 
#> Uncertainty intervals (equal-tailed) and p values (two-tailed) computed using a
#>   Wald z-distribution approximation.
#> ======================